abstract servercall and update layout
