+++ client/src/components/homepage.module.css
@@ -120,18 +120,17 @@
 /* All resource board */
 .resourceBoardContainer {
     position: fixed;
-    z-index: 10;
     top: -1%;
     right: 7%;
-    height: 30rem;
-    width: 30rem;
+    height: 55vh;
+    width: 55vh;
     /*border: 2px dashed yellow;*/
 }
 
 /* Board img bg */
 .resourceBoardBg {
     width: 100%;
-    height: auto;
+    height: 100%;
     display: block;
     filter: drop-shadow(4px 4px 8px rgba(0, 0, 0, 0.3));
 }
@@ -154,11 +153,11 @@
 }
 
 .resourceBoardTitle {
-    font-size: 2.5rem;
+    font-size: 5vh;
     font-weight: 700;
     color: var(--color-brown);
     text-align: center;
-    margin: 2.5rem 0 1.8rem 0;
+    margin: 3.5vh 0 1.8vh 0;
     text-shadow: 1px 1px 2px rgba(255, 255, 255, 0.8);
     text-align: center;
 }
@@ -169,10 +168,10 @@
     align-items: flex-start;
     width:100%;
     gap: 1rem;
-    margin-bottom:2rem;
+    margin-bottom:4vh;
 }
 .resourceIcon {
-    width: clamp(2.5rem, 5%, 3rem);
+    width: 5vh;
     height: auto;
     object-fit: contain;
     flex-shrink: 0;
@@ -204,7 +203,7 @@
 .resourceLabel {
     font-weight: 800;
     color: var(--color-brown);
-    font-size: clamp(1.4rem, 2vw, 1.8rem);
+    font-size: 1.5vw;
     line-height: 1.2;
     text-shadow: 0.5px 0.5px 1px rgba(255, 255, 255, 0.5);
     white-space: nowrap;
@@ -214,7 +213,7 @@
 .resourceValue {
     font-weight: 600;
     color: var(--color-brown);
-    font-size: clamp(1.4rem, 2vw, 1.8rem);
+    font-size: 1.5vw;
     white-space: nowrap;
     flex-shrink: 0;
 
@@ -222,7 +221,7 @@
 
 .progressBarContainer {
     width: 100%;
-    height: 1.2rem;
+    height: 2vh;
     background-color: rgba(255, 255, 255, 0.4);
     border-radius: 12px;
     overflow: hidden;
@@ -263,7 +262,7 @@
 
 
 .homepageWrapper .helloTitle {
-    font-size: 4rem;
+    font-size: 4vw;
     font-weight: 600;
     user-select: none;
     text-align: left;
@@ -274,6 +273,8 @@
     -webkit-text-stroke-color: var(--color-darkGreen);
     -webkit-text-stroke-width: 0.15rem;
     text-shadow: 0 0.5rem 0.5rem rgba(0, 0, 0, 0.25);
+    position:relative;
+    z-index:15;
 }
 
 
@@ -309,24 +310,20 @@
 }
 
 .homepageWrapper .buttonSingle {
-    min-width: 5rem;
-    font-size: 1.5rem;
-    color: black;
+    min-width: 6rem;
+    font-size: 2rem;
+    padding: 1rem 3rem;
 }
 
 
 .homepageWrapper .buttonLogout {
-    padding: 15px 40px;
-    border-radius: 30px;
     position: fixed;
     bottom: clamp(20px, 4vw, 40px);
     right: clamp(20px, 4vw, 40px);
 }
 
 .homepageWrapper .buttonSetting {
-    padding: 15px 40px;
-    border-radius: 30px;
     position: fixed;
     bottom: clamp(20px, 4vw, 40px);
-    right: clamp(180px, 20vw, 220px);
+    right: clamp(200px, 20vw, 240px);
 }
